backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
boto3==1.42.59
botocore==1.42.59
CacheControl==0.14.4
cbor2==5.8.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
clamd==1.0.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cron_descriptor==2.0.6
cryptography==46.0.5
daphne==4.2.1
diff-match-patch==20241021
dj-rest-auth==7.1.1
Django==5.2.11
django-allauth==65.14.3
django-appconf==1.2.0
django-celery-beat==2.8.1
django-clamd==1.0.0
django-constance==4.3.4
django-cors-headers==4.9.0
django-filter==25.2
django-import-export==4.4.0
django-location-field==2.7.3
django-modeltranslation==0.19.19
django-rest-swagger==2.2.0
django-storages==1.14.6
django-tenants==3.10.0
django-timezone-field==7.2.1
django-tinymce==5.0.0
django-unfold==0.81.0
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==40.5.1
fcm-django==2.3.1
fhir.resources==8.2.0
fhir_core==1.1.5
firebase_admin==7.2.0
frozenlist==1.8.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
http_ece==1.2.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
livekit-api==1.1.0
livekit-protocol==1.1.2
MarkupSafe==3.0.3
msgpack==1.1.2
multidict==6.7.1
oauthlib==3.3.1
openapi-codec==1.3.2
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg==3.3.3
py-ubjson==0.16.1
py-vapid==1.9.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pywebpush==2.3.0
PyYAML==6.0.3
redis==7.2.1
referencing==0.37.0
reportlab==4.4.10
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
service-identity==24.2.0
simplejson==3.20.2
six==1.17.0
sqlparse==0.5.5
swapper==1.4.0
tablib==3.9.0
tenant-schemas-celery==4.0.3
Twisted==25.5.0
txaio==25.12.2
types-protobuf==6.32.1.20260221
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
websockets==16.0
yarl==1.22.0
zope.interface==8.2
